edit too much files

+++ archives-manage-entity/src/main/java/com/yintu/rixing/system/SysApprovalProcess.java
@@ -19,7 +19,7 @@ import lombok.EqualsAndHashCode;
 @Data
 @EqualsAndHashCode(callSuper = true)
 @TableName("sys_approval_process")
-@ApiModel(value="SysApprovalProcess对象", description="系统审批流程表")
+@ApiModel(value = "SysApprovalProcess对象", description = "系统审批流程表")
 public class SysApprovalProcess extends BaseEntity {
 
     private static final long serialVersionUID = 1L;
@@ -30,11 +30,11 @@ public class SysApprovalProcess extends BaseEntity {
 
     @ApiModelProperty(value = "审批模式 1.固定 2.灵活")
     @TableField("approval_model")
-    private Integer approvalModel;
+    private Short approvalModel;
 
     @ApiModelProperty(value = "是否审批 1.是 0.否")
     @TableField("approval")
-    private Integer approval;
+    private Short approval;
 
 
 }
